requirements.txt
streamlit==1.33.0
google-generativeai==0.5.2
pandas==2.2.2
plotly==5.22.0
requests==2.32.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
